feature : 폴더 이름 수정 모달 - 기존 모달 사용 및 위치 이동
